move http listeners out from shelob.go
